+++ components/info/api/src/main/java/com/flipperdevices/info/api/screen/InfoScreenProvider.kt
@@ -1,10 +1,11 @@
 package com.flipperdevices.info.api.screen
 
+import com.flipperdevices.deeplink.model.Deeplink
 import com.github.terrakok.cicerone.Screen
 
 /**
  * Provide screens for info components
  */
 interface InfoScreenProvider {
-    fun deviceInformationScreen(): Screen
+    fun deviceInformationScreen(deeplink: Deeplink? = null): Screen
 }