+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/api/InfoFeatureEntryImpl.kt
@@ -1,11 +1,14 @@
 package com.flipperdevices.info.impl.api
 
+import android.content.Intent
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
+import androidx.navigation.navDeepLink
 import androidx.navigation.navigation
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
+import com.flipperdevices.deeplink.model.DeeplinkConstants
 import com.flipperdevices.info.impl.compose.screens.ComposableDeviceInfoScreen
 import com.flipperdevices.info.impl.compose.screens.ComposableFullDeviceInfoScreen
 import com.flipperdevices.settings.api.SettingsFeatureEntry
@@ -20,6 +23,9 @@ class InfoFeatureEntryImpl @Inject constructor(
     private val updaterCardApi: UpdaterCardApi,
     private val settingFeatureEntry: SettingsFeatureEntry
 ) : InfoFeatureEntry {
+
+    private val deeplinkKey = "{${DeeplinkConstants.KEY}}"
+
     private fun start(): String = "@${ROUTE.name}"
 
     override fun fullInfo(): String = "@${ROUTE.name}full"
@@ -27,14 +33,33 @@ class InfoFeatureEntryImpl @Inject constructor(
     override fun NavGraphBuilder.navigation(navController: NavHostController) {
         navigation(startDestination = start(), route = ROUTE.name) {
             composable("@${ROUTE.name}") {
-                ComposableDeviceInfoScreen(updaterCardApi, onOpenFullDeviceInfo = {
-                    navController.navigate(fullInfo())
-                }, onOpenOptions = {
-                        navController.navigate(settingFeatureEntry.ROUTE.name)
-                    })
+                ComposableDeviceInfoScreen(
+                    updaterCardApi,
+                    onOpenFullDeviceInfo = { navController.navigate(fullInfo()) },
+                    onOpenOptions = { navController.navigate(settingFeatureEntry.ROUTE.name) }
+                )
             }
             composable("@${ROUTE.name}full") {
                 ComposableFullDeviceInfoScreen(navController)
+            }
+            composable(
+                route = "@${ROUTE.name}/$deeplinkKey",
+                deepLinks = listOf(
+                    navDeepLink {
+                        uriPattern = "${DeeplinkConstants.LAB_FLIPPER_NET}/$deeplinkKey"
+                        action = Intent.ACTION_VIEW
+                    },
+                    navDeepLink {
+                        uriPattern = "${DeeplinkConstants.MY_FLIPPER_DEV}/$deeplinkKey"
+                        action = Intent.ACTION_VIEW
+                    }
+                )
+            ) {
+                ComposableDeviceInfoScreen(
+                    updaterCardApi,
+                    onOpenFullDeviceInfo = { navController.navigate(fullInfo()) },
+                    onOpenOptions = { navController.navigate(settingFeatureEntry.ROUTE.name) }
+                )
             }
         }
     }