+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [Feature] Add error screen on mfkey32
 - [Feature] Not found if key not found in mfkey32
 - [Feature] Fix format for user keys
+- [Feature] Updater card display update from url
 - [FIX] Options button in device screen
 - [FIX] WearOS connection
 - [FIX] Screen streaming broke